Fix GemDOS Open, it shouldn't truncate like it did with "wb" mode (write only).  Use "rb+" mode instead (read/write), that's the best we can using fopen(). (only open() would allow open write-only without truncating).
Use chmod() in GemDOS Fattrib function so that it can set files
read-only.

Complain in Create and Fattrib if volume label attrib is used
and add some FIXMEs for rest of attribs.

(Change also these functions to bail out on errors instead of using
else clauses, this way they don't need to be indented as much
and it's easier to see the function exit point.)