output what value was set by options if it's not otherwise shown by UI
CPU level etc are shown in statusbar and window size is visible there
too, but for several other options there's no feedback whether Hatari
got correct values.

This is important because their values are typically parsed with
atoi(), but that stops parsing at first unrecognized value
(e.g. 0x2 is interpreted as 0).

I noticed this when testing expression evaluation with debugger
"setopt" command, "1+1" yields "$2" which is interpreted as 0...